Fix Windows MSVC warnings-as-errors and macOS missing Vulkan loader
- include/vkm/renderer/backend/common/render_pass.h: the clear-color/depth
  hash functions passed a bare float to std::hash<uint32_t>(), an implicit
  narrowing conversion. GCC/Clang tolerate this today (the project already
  disables the equivalent warning there), but MSVC's /W4 flags it as C4244
  and /WX escalates it to a hard error - never seen before since Windows
  CI never compiled this far. Made the truncation explicit via
  static_cast<uint32_t>(...), same numeric behavior, no more warning.
- buildscripts/CompileOptions.cmake: disable C4244 (narrowing conversion)
  and C4267 (size_t truncation, hit inside vendored backward-cpp) for
  MSVC, matching the already-tolerated GCC/Clang behavior and the existing
  /wd-suppression pattern in this file for other third-party-triggered
  warnings.
- .github/workflows/macos.yml, scripts/setup-macos.sh: the brew-based
  Vulkan SDK install only added molten-vk/vulkan-headers/vulkan-tools,
  missing vulkan-loader (libvulkan.dylib) - the actual library volk
  dynamically loads at runtime. Without it, volkInitialize() fails and
  VkmDriverVulkan::initializeInner() returns Failed, not caught by the
  HardwareUnsupported skip path (this is a real config bug, not a hardware
  limitation). humbletim/setup-vulkan-sdk provided this implicitly via its
  "Vulkan-Loader" component; the replacement brew install needs it too.
- .github/workflows/ubuntu.yml, scripts/setup-linux.sh: add libwayland-dev
  (the actual Wayland client library pkg-config needs) alongside the
  already-added libwayland-bin/wayland-protocols (wayland-scanner) -
  GLFW's Wayland pkg_check_modules(REQUIRED) needs both.

+++ include/vkm/renderer/backend/common/render_pass.h
@@ -120,10 +120,10 @@ struct std::hash<vkm::VkmColorAttachmentDescriptor>
         return std::hash<uint32_t>()(descriptor._attachmentId) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._loadAction)) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._storeAction)) ^
-               std::hash<uint32_t>()(descriptor._clearColors[0]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[1]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[2]) ^
-               std::hash<uint32_t>()(descriptor._clearColors[3]);
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[0])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[1])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[2])) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearColors[3]));
     }
 };
 
@@ -135,7 +135,7 @@ struct std::hash<vkm::VkmDepthStencilAttachmentDescriptor>
         return std::hash<uint32_t>()(descriptor._attachmentId) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._loadAction)) ^
                std::hash<uint8_t>()(static_cast<uint8_t>(descriptor._storeAction)) ^
-               std::hash<uint32_t>()(descriptor._clearDepth) ^
+               std::hash<uint32_t>()(static_cast<uint32_t>(descriptor._clearDepth)) ^
                std::hash<uint32_t>()(descriptor._clearStencil);
     }
 };
